feat: refactor QueryEditor to use custom hooks for selected query and datasource management
